+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/list/viewmodel/ChannelListViewModel.kt
@@ -6,7 +6,6 @@ import androidx.lifecycle.Transformations
 import androidx.lifecycle.Transformations.map
 import androidx.lifecycle.ViewModel
 import com.getstream.sdk.chat.utils.extensions.isDraft
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Filters
@@ -20,7 +19,7 @@ import io.getstream.chat.android.livedata.controller.QueryChannelsController
 /**
  * ViewModel class for [io.getstream.chat.android.ui.channel.list.ChannelListView].
  * Responsible for keeping the channels list up to date.
- * Can be bound to the view using [ChannelsViewModel.bindView] function.
+ * Can be bound to the view using [ChannelListViewModel.bindView] function.
  * @param chatDomain entry point for all livedata & offline operations
  * @param filter filter for querying channels, should never be empty
  * @param sort defines the ordering of the channels
@@ -34,7 +33,7 @@ public class ChannelListViewModel(
         Filters.ne("draft", true)
     ),
     private val sort: QuerySort<Channel> = DEFAULT_SORT,
-    private val limit: Int = 30
+    private val limit: Int = 30,
 ) : ViewModel() {
     private val stateMerger = MediatorLiveData<State>()
     public val state: LiveData<State> = stateMerger
@@ -45,18 +44,25 @@ public class ChannelListViewModel(
     public val paginationState: LiveData<PaginationState> = Transformations.distinctUntilChanged(paginationStateMerger)
 
     init {
-        stateMerger.value = State.Loading
+        stateMerger.value = INITIAL_STATE
         chatDomain.useCases.queryChannels(filter, sort, limit).enqueue { queryChannelsControllerResult ->
+            val currentState = stateMerger.value!!
             if (queryChannelsControllerResult.isSuccess) {
                 val queryChannelsController = queryChannelsControllerResult.data()
                 stateMerger.addSource(
                     map(queryChannelsController.channelsState) { channelState ->
                         when (channelState) {
                             is QueryChannelsController.ChannelsState.NoQueryActive,
-                            is QueryChannelsController.ChannelsState.Loading -> State.Loading
-                            is QueryChannelsController.ChannelsState.OfflineNoResults -> State.NoChannelsAvailable
-                            is QueryChannelsController.ChannelsState.Result ->
-                                State.Result(channelState.channels.filterNot { it.hidden == true || it.isDraft })
+                            is QueryChannelsController.ChannelsState.Loading,
+                            -> currentState.copy(isLoading = true)
+                            is QueryChannelsController.ChannelsState.OfflineNoResults -> currentState.copy(
+                                isLoading = false,
+                                channels = emptyList(),
+                            )
+                            is QueryChannelsController.ChannelsState.Result -> currentState.copy(
+                                isLoading = false,
+                                channels = channelState.channels.filterNot { it.hidden == true || it.isDraft },
+                            )
                         }
                     }
                 ) { state -> stateMerger.value = state }
@@ -67,7 +73,10 @@ public class ChannelListViewModel(
                     setPaginationState { copy(endOfChannels = endOfChannels) }
                 }
             } else {
-                stateMerger.postValue(State.NoChannelsAvailable)
+                stateMerger.value = currentState.copy(
+                    isLoading = false,
+                    channels = emptyList(),
+                )
             }
         }
     }
@@ -75,10 +84,6 @@ public class ChannelListViewModel(
     public fun onAction(action: Action) {
         when (action) {
             is Action.ReachedEndOfList -> requestMoreChannels()
-            is Action.LogoutClicked -> {
-                ChatClient.instance().disconnect()
-                stateMerger.postValue(State.NavigateToLoginScreen)
-            }
         }.exhaustive
     }
 
@@ -106,25 +111,21 @@ public class ChannelListViewModel(
         paginationStateMerger.value = reducer(paginationStateMerger.value ?: PaginationState())
     }
 
-    public sealed class State {
-        public object Loading : State()
-        public data class Result(val channels: List<Channel>) : State()
-        public object NoChannelsAvailable : State()
-        public object NavigateToLoginScreen : State()
-    }
+    public data class State(val isLoading: Boolean, val channels: List<Channel>)
 
     public data class PaginationState(
         val loadingMore: Boolean = false,
-        val endOfChannels: Boolean = false
+        val endOfChannels: Boolean = false,
     )
 
     public sealed class Action {
         public object ReachedEndOfList : Action()
-        public object LogoutClicked : Action()
     }
 
     public companion object {
         @JvmField
         public val DEFAULT_SORT: QuerySort<Channel> = QuerySort.desc("last_updated")
+
+        private val INITIAL_STATE: State = State(isLoading = true, channels = emptyList())
     }
 }